move spack.llnl.util.link_tree to spack.util.link_tree
