+++ lib/spack/spack/cmd/view.py
@@ -43,8 +43,8 @@ import spack.llnl.util.tty as tty
 import spack.schema.projections
 import spack.store
 from spack.config import validate
-from spack.llnl.util.link_tree import MergeConflictError
 from spack.util import spack_yaml as s_yaml
+from spack.util.link_tree import MergeConflictError
 
 description = "manipulate view directories in the filesystem"
 section = "environments"
