+++ lib/spack/spack/environment/environment.py
@@ -57,10 +57,10 @@ from spack import traverse
 from spack.enums import ConfigScopePriority
 from spack.llnl.util.filesystem import copy_tree, islink, readlink
 from spack.llnl.util.lang import stable_partition
-from spack.llnl.util.link_tree import ConflictingSpecsError
 from spack.schema.env import TOP_LEVEL_KEY
 from spack.spec import Spec
 from spack.spec_filter import SpecFilter
+from spack.util.link_tree import ConflictingSpecsError
 from spack.util.path import substitute_path_variables
 
 from .list import SpecList, SpecListError, SpecListParser
