+++ lib/spack/spack/util/link_tree.py
